requirements.txt
openai==0.28.1
flask~=3.0.0
werkzeug~=3.0.1
Pillow~=10.0.1